chore(app): Paginate the advanced search on the backend
This should drastically improve the performance with large databases.
