+++ app/config/packages/doctrine.yaml
@@ -1,7 +1,7 @@
 doctrine:
     dbal:
         url: '%env(resolve:DATABASE_URL)%'
-        schema_filter: '~^(?!out_|mail_count|sql_limit_report)~'
+        schema_filter: '~^(?!out_quarantine|out_wblist|mail_count|sql_limit_report)~'
 
         # IMPORTANT: You MUST configure your server version,
         # either here or in the DATABASE_URL env var (see .env file)
