+++ app/src/Controller/SearchController.php
@@ -5,117 +5,72 @@ namespace App\Controller;
 use App\Entity\Msgs;
 use App\Entity\User;
 use App\Form\SearchFilterType;
+use App\Repository\MsgrcptSearchRepository;
+use App\Repository\OutMsgrcptRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
-use Symfony\Component\HttpFoundation\JsonResponse;
 
 #[IsGranted('ROLE_ADMIN')]
 #[Route(path: '/advanced_search')]
 class SearchController extends AbstractController
 {
-    public function __construct(private EntityManagerInterface $em)
-    {
+    public function __construct(
+        private EntityManagerInterface $em,
+        private MsgrcptSearchRepository $msgrcptSearchRepository,
+        private OutMsgrcptRepository $outMsgrcptRepository
+    ) {
     }
 
     #[Route(path: '/', name: 'advanced_search', methods: ['GET', 'POST'])]
-    public function advancedSearch(Request $request): Response
-    {
-        $form = $this->createForm(SearchFilterType::class);
+    public function advancedSearch(
+        Request $request,
+        PaginatorInterface $paginator
+    ): Response {
+        $form = $this->createForm(SearchFilterType::class, [], ['method' => 'GET']);
         $form->handleRequest($request);
 
-        $data = $form->getData();
-        $messageType = $data['messageType'] ?? 'incoming';
+        $this->sanitizeSortParameter($request);
 
-        $sortParams = null;
-        if ($request->request->has('sortField') && $request->request->has('sortDirection')) {
-            $sortParams = [
-                'sort' => $request->request->getString('sortField'),
-                'direction' => $request->request->getString('sortDirection'),
-            ];
-        }
-
-        /** @var User $user */
-        $user = $this->getUser();
-        $allMessages = $this->em->getRepository(Msgs::class)->advancedSearch($user, $messageType, $sortParams);
-
-
-        // Initialize active filters
         $activeFilters = [];
 
-        // If form is submitted and valid, set active filters and filter messages
-        if ($form->isSubmitted() && $form->isValid()) {
-            $data = $form->getData();
+        $activeFilters = $form->getData();
 
-            // Track which filters are active
-            foreach ($data as $key => $value) {
-                if (!empty($value)) {
-                    $activeFilters[$key] = true;
-                }
-            }
-
-            // Apply all active filters to the messages
-            $allMessages = array_filter($allMessages, function ($message) use ($data) {
-                // Basic filters
-                if (!empty($data['fromAddr']) && stripos($message['from_addr'], $data['fromAddr']) === false) {
-                    return false;
-                }
-                if (!empty($data['email']) && stripos($message['email'], $data['email']) === false) {
-                    return false;
-                }
-                if (!empty($data['subject']) && stripos($message['subject'], $data['subject']) === false) {
-                    return false;
-                }
-                if (!empty($data['mailId']) && stripos($message['mail_id'], $data['mailId']) === false) {
-                    return false;
-                }
-
-                // Advanced filters
-                if ($data['bspamLevelMin'] !== null && $message['bspam_level'] < $data['bspamLevelMin']) {
-                    return false;
-                }
-                if ($data['bspamLevelMax'] !== null && $message['bspam_level'] > $data['bspamLevelMax']) {
-                    return false;
-                }
-                if (!empty($data['startDate']) && $message['time_iso'] < $data['startDate']->format('Ymd\THis\Z')) {
-                    return false;
-                }
-                if (!empty($data['endDate']) && $message['time_iso'] > $data['endDate']->format('Ymd\THis\Z')) {
-                    return false;
-                }
-                if (!empty($data['size']) && stripos($message['size'], $data['size']) === false) {
-                    return false;
-                }
-                if (!empty($data['host']) && stripos($message['host'], $data['host']) === false) {
-                    return false;
-                }
-                if (!empty($data['replyTo']) && $message['replyTo'] !== $data['replyTo']) {
-                    return false;
-                }
-                return true;
-            });
+        $messageType = $activeFilters['messageType'] ?? 'incoming';
+        if ($messageType == 'incoming') {
+            $allMessages = $this->msgrcptSearchRepository
+                ->getAdvancedSearchQuery($activeFilters);
+        } else {
+            $allMessages = $this->outMsgrcptRepository
+                ->getAdvancedSearchQuery($activeFilters);
         }
 
-        // Handle AJAX requests
-        if ($request->isXmlHttpRequest()) {
-            return new JsonResponse([
-                'content' => $this->renderView('search/_messages.html.twig', [
-                    'msgs' => $allMessages,
-                    'messageType' => $messageType,
-                    'activeFilters' => $activeFilters,
-                ]),
-            ]);
-        }
+        $perPage = (int) $this->getParameter('app.per_page_global');
+        $perPage = $request->getSession()->has('perPage') ? $request->getSession()->get('perPage') : $perPage;
 
-        // Render the main template
+        $recipients = $paginator->paginate(
+            $allMessages,
+            $request->query->getInt('page', 1),
+            $perPage,
+            [
+                'wrap-queries' => true,
+                'fetchJoinCollection' => false,
+                'distinct' => false,
+                'defaultSortFieldName' => 'm.timeIso',
+                'defaultSortDirection' => 'desc',
+            ]
+        );
+
+
         return $this->render('search/advanced_search.html.twig', [
-            'msgs' => $allMessages,
+            'messagesRecipients' => $recipients,
             'form' => $form->createView(),
             'messageType' => $messageType,
-            'activeFilters' => $activeFilters, // Pass the activeFilters to the main view
+            'activeFilters' => $activeFilters,
         ]);
     }
 
@@ -168,5 +123,25 @@ class SearchController extends AbstractController
             'outMsgRcpt' => $outMsgRcpt,
             'message' => $ms,
         ]);
+    }
+
+    private function sanitizeSortParameter(Request $request): void
+    {
+        $allowedSortFields = [
+            'mr.mailId',
+            'm.fromAddr',
+            'maddr.email',
+            'm.subject',
+            'm.timeIso',
+            'mr.bspamLevel',
+            'm.size',
+            'm.host',
+        ];
+
+        $sort = $request->query->getString('sort', 'm.timeIso');
+
+        if (!in_array($sort, $allowedSortFields, true)) {
+            $request->query->set('sort', 'm.timeIso');
+        }
     }
 }
